ref : 스케줄 조회 api 구조 재배치 및 리팩터링
