+++ homecare/src/main/java/jaega/homecare/domain/center/controller/CenterController.java
@@ -3,10 +3,11 @@ package jaega.homecare.domain.center.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jaega.homecare.domain.WorkMatch.dto.res.GetCaregiverMatchesByMonth;
+import jaega.homecare.domain.WorkMatch.dto.res.GetCaregiverMatchesResponse;
 import jaega.homecare.domain.center.dto.req.CreateCaregiverRequest;
 import jaega.homecare.domain.center.dto.res.GetCaregiverResponse;
-import jaega.homecare.domain.center.dto.res.GetMatchingResultResponse;
-import jaega.homecare.domain.serviceMatch.dto.res.ServiceMatchNotificationResponse;
+import jaega.homecare.domain.serviceMatch.dto.res.GetServiceMatchByCenterResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -31,5 +32,18 @@ public interface CenterController {
     @Operation(summary = "배정 내역 전체 조회 API", description = "배정된 신청자-요양보호사 전체 목록을 최신순으로 조회합니다.")
     @ApiResponse(responseCode = "200", description = "요양 보호사 전체 조회 성공")
     @GetMapping("/{centerId}/assign")
-    public ResponseEntity<List<ServiceMatchNotificationResponse>> getAllMatchingResult(@PathVariable UUID centerId);
+    public ResponseEntity<List<GetServiceMatchByCenterResponse>> getAllMatchingResult(@PathVariable UUID centerId);
+
+    @Operation(summary = "특정 요양 보호사의 매칭 스케줄 조회", description = "특정 요양 보호사의 매칭 스케줄들을 조회합니다.")
+    @ApiResponse(responseCode = "200", description = "특정 요양 보호사의 매칭 스케줄 조회 성공")
+    @GetMapping("/schedule/{caregiverId}")
+    ResponseEntity<List<GetCaregiverMatchesResponse>> getWorkMatchByCaregiver(@PathVariable UUID caregiverId);
+
+    @Operation(summary = "특정 년도, 월의 요양보호사 매칭 스케줄 조회", description = "특정 년도, 월에 해당하는 요양보호사의 스케줄을 조회합니다.")
+    @ApiResponse(responseCode = "200", description = "특정 년도, 월의 요양보호사 매칭 스케줄 조회")
+    @GetMapping("/schedule/date")
+    ResponseEntity<List<GetCaregiverMatchesByMonth>> getMatchesByMonth(
+            @RequestParam int year,
+            @RequestParam int month
+    );
 }
