+++ homecare/src/main/java/jaega/homecare/domain/center/controller/CenterControllerImpl.java
@@ -1,23 +1,22 @@
 package jaega.homecare.domain.center.controller;
 
+import jaega.homecare.domain.WorkMatch.dto.res.GetCaregiverMatchesByMonth;
+import jaega.homecare.domain.WorkMatch.dto.res.GetCaregiverMatchesResponse;
+import jaega.homecare.domain.WorkMatch.service.query.WorkMatchQueryService;
 import jaega.homecare.domain.caregiver.service.command.CaregiverCommandService;
 import jaega.homecare.domain.caregiver.service.query.CaregiverQueryService;
 import jaega.homecare.domain.center.dto.req.CreateCaregiverRequest;
 import jaega.homecare.domain.center.dto.res.GetCaregiverResponse;
-import jaega.homecare.domain.center.dto.res.GetMatchingResultResponse;
 import jaega.homecare.domain.center.entity.Center;
 import jaega.homecare.domain.center.service.command.CenterCommandService;
 import jaega.homecare.domain.center.service.query.CenterQueryService;
-import jaega.homecare.domain.serviceMatch.dto.res.ServiceMatchNotificationResponse;
+import jaega.homecare.domain.serviceMatch.dto.res.GetServiceMatchByCenterResponse;
 import jaega.homecare.domain.serviceMatch.service.query.ServiceMatchQueryService;
 import jaega.homecare.domain.users.entity.User;
 import jaega.homecare.domain.users.entity.UserRole;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.UUID;
@@ -32,6 +31,7 @@ public class CenterControllerImpl implements CenterController{
     private final CaregiverCommandService caregiverCommandService;
     private final CaregiverQueryService caregiverQueryService;
     private final ServiceMatchQueryService serviceMatchQueryService;
+    private final WorkMatchQueryService workMatchQueryService;
 
     @Override
     public ResponseEntity<Void> createCaregiver(@RequestBody CreateCaregiverRequest createCaregiverRequest, @PathVariable UUID centerId){
@@ -49,8 +49,23 @@ public class CenterControllerImpl implements CenterController{
     }
 
     @Override
-    public ResponseEntity<List<ServiceMatchNotificationResponse>> getAllMatchingResult(@PathVariable UUID centerId){
-        List<ServiceMatchNotificationResponse> notifications = serviceMatchQueryService.getMatchesByCenter(centerId);
+    public ResponseEntity<List<GetServiceMatchByCenterResponse>> getAllMatchingResult(@PathVariable UUID centerId){
+        List<GetServiceMatchByCenterResponse> notifications = serviceMatchQueryService.getMatchesByCenter(centerId);
         return ResponseEntity.ok(notifications);
+    }
+
+    @Override
+    public ResponseEntity<List<GetCaregiverMatchesResponse>> getWorkMatchByCaregiver(@PathVariable UUID caregiverId) {
+        List<GetCaregiverMatchesResponse> responses = workMatchQueryService.getWorkMatchesByCaregiver(caregiverId);
+        return ResponseEntity.ok(responses);
+    }
+
+    @Override
+    public ResponseEntity<List<GetCaregiverMatchesByMonth>> getMatchesByMonth(
+            @RequestParam int year,
+            @RequestParam int month
+    ) {
+        List<GetCaregiverMatchesByMonth> response = workMatchQueryService.getWorkMatchesByMonth(year, month);
+        return ResponseEntity.ok(response);
     }
 }
