+++ homecare/src/main/java/jaega/homecare/domain/consumer/controller/ConsumerControllerImpl.java
@@ -5,14 +5,20 @@ import jaega.homecare.domain.WorkMatch.service.command.WorkMatchCommandService;
 import jaega.homecare.domain.consumer.dto.req.ConfirmCaregiverRequest;
 import jaega.homecare.domain.consumer.dto.req.ConsumerCreateRequest;
 import jaega.homecare.domain.serviceMatch.dto.req.CreateServiceMatchRequest;
+import jaega.homecare.domain.serviceMatch.dto.res.GetServiceMatchByConsumerResponse;
 import jaega.homecare.domain.serviceMatch.service.command.ServiceMatchCommandService;
+import jaega.homecare.domain.serviceMatch.service.query.ServiceMatchQueryService;
 import jaega.homecare.domain.users.entity.UserRole;
 import jaega.homecare.domain.consumer.service.command.ConsumerCommandService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+import java.util.UUID;
 
 @RequiredArgsConstructor
 @RestController
@@ -22,6 +28,7 @@ public class ConsumerControllerImpl implements ConsumerController {
     private final ConsumerCommandService consumerCommandService;
     private final ServiceMatchCommandService serviceMatchCommandService;
     private final WorkMatchCommandService workMatchCommandService;
+    private final ServiceMatchQueryService serviceMatchQueryService;
 
     @Override
     public ResponseEntity<Void> createConsumer(@RequestBody ConsumerCreateRequest request) {
@@ -48,6 +55,12 @@ public class ConsumerControllerImpl implements ConsumerController {
         workMatchCommandService.createWorkMatch(createWorkMatchRequest);
 
         return ResponseEntity.noContent().build();
+    }
+
+    @Override
+    public ResponseEntity<List<GetServiceMatchByConsumerResponse>> getMatchesByConsumer(@PathVariable UUID consumerId){
+        List<GetServiceMatchByConsumerResponse> responses = serviceMatchQueryService.getMatchesByConsumer(consumerId);
+        return ResponseEntity.ok(responses);
     }
 
 }
