--- homecare/src/main/java/jaega/homecare/domain/serviceMatch/controller/ServiceMatchController.java
@@ -1,19 +1,0 @@
-package jaega.homecare.domain.serviceMatch.controller;
-
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.tags.Tag;
-import jaega.homecare.domain.serviceMatch.dto.res.ServiceMatchNotificationResponse;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-import java.util.List;
-import java.util.UUID;
-
-@Tag(name = "ServiceMatch", description = "수요자 매칭 API")
-@RequestMapping("/api/consumerSchedule")
-public interface ServiceMatchController {
-
-}
